requirements.txt
discord.py
beautifulsoup4
requests
schedule
psutil
python-dotenv